+++ books/README.md
@@ -1,0 +1,3 @@
+# Books
+
+This directory contains summaries for some books about the compilers and their architectures.
