+++ books/modern-complier-implementation/introduction-to-compilers/introduction.md
@@ -1,0 +1,41 @@
+<img src="" width=40 height=80/>
+# Introduction to the modern compiler implementation in java
+
+This is a condensed version for the the book "Modern compiler implementaion in java 2nd e by Andrew W. Appel - Cambridge-university-press".
+
+Chapters implemented here by this repository:
+- [x] [Lexical Analysis]()
+- [ ] [Parsing]()
+- [ ] [Abstract Syntax]()
+- [ ] [Semantic Analysis]()
+- [ ] [Activation Records]()
+- [ ] [Translation into intermediate code]()
+- [ ] [Basic blocks and traces]()
+- [ ] [Instruction Selection]()
+- [ ] [Liveness Analysis]()
+- [ ] [Register Allocation]()
+
+Compilers adopt a universal model, which modularizes code in `compiler modules` and `compiler inferfaces` aka. Compiler Phases:
+
+![]()
+
+Compiler phases: 
+- [x] Lexing Phase: breaks down the source file into individual words or tokens.
+- [ ] Parsing Phase: analyzes the phrase structure of a program.
+- [ ] Semantic Actions: builds a piece of *abstract syntax tree* corresponding to each phrase.
+- [ ] Semantic Analysis
+- [ ] Frame Layout
+- [ ] Translate
+- [ ] Canonicalize
+- [ ] Instruction Selection
+- [ ] Control flow analysis 
+- [ ] Data flow analysis
+- [ ] Register Allocation
+- [ ] Code Emission
+
+Compiler modules:
+
+Compiler interfaces: Abstract data structures 
+
+
+
