--- lib/src/main/java/com/compiler/grammer/rule/statement/PrintStatement.java
@@ -1,13 +1,0 @@
-package com.compiler.grammer.rule.statement;
-
-import com.compiler.grammer.symbol.ExpressionList;
-import com.compiler.grammer.symbol.Statement;
-
-public class PrintStatement extends Statement {
-    
-    private final ExpressionList expressionList;
-
-    public PrintStatement(final ExpressionList expressionList) {
-        this.expressionList = expressionList;
-    }
-}
