--- lib/src/main/java/com/compiler/grammer/symbol/Statement.java
@@ -1,5 +1,0 @@
-package com.compiler.grammer.symbol;
-
-public abstract class Statement {
-    
-}
