+++ settings.gradle
@@ -8,5 +8,5 @@
  * This project uses @Incubating APIs which are subject to change.
  */
 
-rootProject.name = 'Minijava'
-include('lib')
+rootProject.name = 'Compilers-At-A-Glance'
+include('simple-interpreter-prototype')
