+++ simple-interpreter-prototype/src/main/java/com/interpreter/grammer/rule/expression/ExtendedSequence.java
@@ -1,0 +1,15 @@
+package com.interpreter.grammer.rule.expression;
+
+import com.interpreter.grammer.symbol.Expression;
+import com.interpreter.grammer.symbol.Statement;
+
+public class ExtendedSequence extends Expression {
+
+    private final Statement statement;
+    private final Expression expression;
+
+    public ExtendedSequence(final Statement statement, final Expression expression) {
+        this.statement = statement;
+        this.expression = expression;
+    }
+}
