+++ simple-interpreter-prototype/src/main/java/com/interpreter/grammer/rule/expression/IdentifierExpression.java
@@ -1,0 +1,11 @@
+package com.interpreter.grammer.rule.expression;
+
+import com.interpreter.grammer.symbol.Expression;
+
+public class IdentifierExpression extends Expression {
+    private final String identifier;
+
+    public IdentifierExpression(final String identifier) {
+        this.identifier = identifier;
+    }
+}
