+++ simple-interpreter-prototype/src/main/java/com/interpreter/grammer/rule/expression/NumberExpression.java
@@ -1,0 +1,11 @@
+package com.interpreter.grammer.rule.expression;
+
+import com.interpreter.grammer.symbol.Expression;
+
+public class NumberExpression extends Expression {
+    private final int number;
+
+    public NumberExpression(final int number) {
+        this.number = number;
+    }
+}
