+++ simple-interpreter-prototype/src/main/java/com/interpreter/grammer/symbol/Expression.java
@@ -1,0 +1,5 @@
+package com.interpreter.grammer.symbol;
+
+public abstract class Expression {
+    
+}
