+++ simple-interpreter-prototype/src/main/java/com/interpreter/grammer/symbol/Statement.java
@@ -1,0 +1,5 @@
+package com.interpreter.grammer.symbol;
+
+public abstract class Statement {
+    
+}
