+++ simple-interpreter-prototype/src/test/java/compiler/LibraryTest.java
